+++ app/src/main/java/de/baumann/hhsmoodle/fragmentsMain/FragmentInfo.java
@@ -5,6 +5,7 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.TypedArray;
+import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.design.widget.Snackbar;
@@ -111,7 +112,8 @@ public class FragmentInfo extends Fragment {
                 final CharSequence[] options = {
                         getString(R.string.bookmark_edit_fav),
                         getString(R.string.bookmark_createNote),
-                        getString(R.string.bookmark_CreateNotification)};
+                        getString(R.string.bookmark_createNotification),
+                        getString(R.string.bookmark_createShortcut)};
                 new AlertDialog.Builder(getActivity())
                         .setItems(options, new DialogInterface.OnClickListener() {
                             @Override
@@ -138,7 +140,7 @@ public class FragmentInfo extends Fragment {
                                     startActivity(intent_in);
                                 }
 
-                                if (options[item].equals (getString(R.string.bookmark_CreateNotification))) {
+                                if (options[item].equals (getString(R.string.bookmark_createNotification))) {
                                     SharedPreferences sharedPref = PreferenceManager.getDefaultSharedPreferences(getActivity());
                                     sharedPref.edit()
                                             .putString("noteTitle", title)
@@ -148,6 +150,22 @@ public class FragmentInfo extends Fragment {
                                     Intent intent_in = new Intent(getActivity(), Notes_MainActivity.class);
                                     startActivity(intent_in);
                                 }
+
+                                if (options[item].equals (getString(R.string.bookmark_createShortcut))) {
+
+                                    Intent i = new Intent();
+                                    i.setAction(Intent.ACTION_VIEW);
+                                    i.setData(Uri.parse(url));
+
+                                    Intent shortcut = new Intent();
+                                    shortcut.putExtra("android.intent.extra.shortcut.INTENT", i);
+                                    shortcut.putExtra("android.intent.extra.shortcut.NAME", "THE NAME OF SHORTCUT TO BE SHOWN");
+                                    shortcut.putExtra(Intent.EXTRA_SHORTCUT_NAME, title);
+                                    shortcut.putExtra(Intent.EXTRA_SHORTCUT_ICON_RESOURCE, Intent.ShortcutIconResource.fromContext(getActivity().getApplicationContext(), R.mipmap.ic_launcher));
+                                    shortcut.setAction("com.android.launcher.action.INSTALL_SHORTCUT");
+                                    getActivity().sendBroadcast(shortcut);
+                                }
+
 
                             }
                         }).show();